Avoid fetching event timestamps from Tk
It seems using '%t' doesn't always produce what we want.

Sometimes the time is correct (Windows, Linux),
but on MacOS it's broken (it's a negative number).

For now we can just timestamp the event with the current time
which should be close to the actual time of when the
event was triggered.

Also removed getTclTimestamp() since we can't rely on it.

Fixes #44